worker/requirements.txt
redis==5.1.0
httpx==0.27.2
pyyaml==6.0.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic-settings==2.5.2
